The textures finally work. Hallelujah!!! There are some concurrency issues which can cause it to lock up when loading. The milliseconds texturing has serious issues.
But it works!! =)

Version bump (finally haha): 0.8.37 -> 0.9.42

+++ ChroBars/src/com/psoft/chrobars/opengl/ChroTexture.java
@@ -21,7 +21,7 @@ import com.psoft.chrobars.ChroType;
 public class ChroTexture {
 
 	private boolean cacheLater;
-	private short orderIndex;
+	private int orderIndex;
 	private int resId, texId;
 	private Bitmap bmpTex;
 	private String resName;
@@ -194,7 +194,7 @@ public class ChroTexture {
 	/**
 	 * @return the orderIndex
 	 */
-	public short getOrderIndex() {
+	public int getOrderIndex() {
 		return orderIndex;
 	}
